Enhancements: Particles and New Background
JPong.js: added particle object {lines 14 to 68}. Declared variable
"particleNumber" which defines the number of particles emitted when the
ball collides with the wall. This is a number that can later be
quantified in the menu screen or set to off (zero particles). According
to my test, particles show no performance impact until I use hundreds of
them. I've only found 6 to be necessary to create the effect I want.
Index.html: Changed the styling of the footer.
JPong.css: Added gradient background to Header. Added box-shadowing to
the content. Changed the font of the subfamily to something more
readable.